Separate seat component and create upper status msg for seat state
